+++ CHANGELOG.md
@@ -179,3 +179,8 @@
 - Overhaul for pooling remnants item along with other grabbable objects/items. To increase compatibility with other mods.
 - Updated feature, now you can spawn 0 remnant items on a moon rather than the minimum of 1.
 - Fixed issue: end of round report now shows accurate info regarding scrap items collected and amount of scrap value was on that moon.
+
+#1.4.0
+- Thanks to the help of [Zaggy1024](https://github.com/Zaggy1024), the despawn issues are fixed and it increased the compatibility with others mods drastrictly. 
+- Fixed beltbag bug, now it properly works. (can be changed in the config)
+- The end of round patch gives accurate total scrap collected. (can be changed in the config)